### 2024-01-18 #156 @ghiggi
`disdrodb` review
Submitting Author: Gionata Ghiggi (@ghiggi)
Package Name: disdrodb
One-Line Description of Package: disdrodb -  A software for the decentralized archiving and standardization of global disdrometer data
Repository Link:  https://github.com/ltelab/disdrodb
Version submitted:  v.0.0.21
Version accepted: TBD

---

The raindrop size distribution (DSD) describes the concentration and size distributions of raindrops in a volume of air. It is a crucial piece of information to model the propagation of microwave signals through the atmosphere (key for telecommunication and weather radar remote sensing calibration), to improve microphysical schemes in numerical weather prediction models, and to understand land surface processes (rainfall interception, soil erosion).

Recognizing the importance of understanding DSD's spatial and temporal variability, scientists worldwide have initiated efforts to "count the drops" by deploying disdrometers—specialized instruments designed to record DSD. Numerous measurement campaigns have been conducted by meteorological services, national agencies (e.g., NASA, ARM, NCAR), and university research groups. Despite these efforts, a significant portion of the collected data remains difficult to access. These data are often stored in diverse formats with inadequate documentation and metadata, posing challenges in sharing, analyzing, comparing, and reusing the data.

In response to these challenges, the _disdrodb_ Python package offers:

1.  A Decentralized Data Archive Infrastructure: The _disdrodb_ package establishes a decentralized data archive, fostering the exchange and retrieval of raw disdrometer data within the scientific community. This infrastructure addresses the issue of data accessibility, documentation and promotes collaborative research.

2. Standardization of Raw Data. The _disdrodb_ package provides tools to convert heterogeneous raw data into a uniform netCDF4 format, known as the DISDRODB L0 product. This standardization is a significant step forward, ensuring that data from different sources become compatible and easier to analyze, compare, and share, thereby enhancing the overall utility and reusability of the data.

**How the and why the package falls under the categories you indicated above**

**Data Retrieval**

The disdrodb package facilitates the retrieval of raw measurement acquired by disdrometer stations which are included in the DISDRODB Decentralized Data Archive. This remote archive  comprises public cloud repositories such as [Zenodo](https://zenodo.org/communities/disdrodb?q=&l=list&p=1&s=10&sort=newest). The disdrodb package tracks the available stations through the [DISDRODB Metadata Archive which is hosted on GitHub](https://github.com/ltelab/disdrodb-data).

**Data Munging**

After downloading the desired data, users can use disdrodb to convert the heterogeneous raw data into a uniform netCDF4 format (DISDRODB L0) with a single command. This conversion facilitates subsequent scientific analysis and product generation. For each disdrometer station, the disdrodb python package has a specialized reader that enable to accurately parse the raw sensor data.

**Data Deposition**

The disdrodb package offers a workflow for users who wish to contribute their disdrometer measurements to the DISDRODB community. This workflow ensures the long-term documentation of the data and simplifies the data upload process to the DISDRODB Decentralized Data Archive. Users must perform three main tasks:

- Create a reader that reads the raw data into a dataframe, adhering to the DISDRODB guidelines.

- Provide the metadata of the disdrometer station, which will be included in the DISDRODB Metadata Archive.

- Upload the station's raw data to a remote repository and insert the station data URL into the DISDRODB Metadata Repository. The disdrodb package can automate this final step if the chosen remote repository is [Zenodo](https://zenodo.org/communities/disdrodb?q=&l=list&p=1&s=10&sort=newest).

**Who is the target audience and what are scientific applications of this package?**

The primary audience for this package includes researchers and students in the fields of remote sensing and atmospheric science, specifically those focused on precipitation. The package is designed to support applications in remote sensing and atmospheric science.

**Are there other Python packages that accomplish the same thing? If so, how does yours differ?**

To our knowledge, there are no other packages that offer an integrated infrastructure for retrieving, sharing, archiving, reading, and standardizing disdrometer data.

However, the [pyDSD package](https://github.com/josephhardinee/PyDSD) exists for studying the DSD. It provides methods for high-level scientific analysis of disdrometer raw data, such as computing DSD parameters and simulating weather radar reflectivities.

The DISDRODB Working Group plans to leverage and adapt pyDSD codes in the future to generate uniform, high-level scientific products for all stations within the DISDRODB Global Archive.
