### 2024-02-05 b25c069
Improve documentation compilation
