+++ docs/source/maintainers_guidelines.rst
@@ -45,9 +45,9 @@ Examples of non-breaking changes include :
 - Fix a bug.
 - Adding the configuration files to process a new sensor.
 - Adding a new DISDRODB L0 reader.
-- Adding new functionalities to the DISDRODB API that don’t affect the behavior of the API directly.
+- Adding new functionalities to the DISDRODB API that do not affect the behavior of the API directly.
 - Updating the documentation.
-- Internal function refactoring that doesn’t affect the behavior of the software directly.
+- Internal function refactoring that does not affect the behavior of the software directly.
 
 
 One implication of this policy is that clients should be prepared to ignore the presence of unexpected fields in responses and unexpected values for enums. This is necessary to preserve compatibility between PATCH versions within the same MINOR version range, since optional fields and enum values can be added as non-breaking changes.
@@ -63,38 +63,67 @@ The `LTE <https://www.epfl.ch/labs/lte/>`_ does not guarantee any period of supp
 Refer to the list of Recommended Releases to see current releases and more details.
 
 
-
-
 Documentation pipeline
 ========================
 
-DISDRODB’s documentation is built using Sphinx. All documentation lives in the ``docs/`` directory of the project repository.
+DISDRODB's documentation is built using the powerful `Sphinx <https://www.sphinx-doc.org/en/master/>`_ framework,
+styled with `Book Theme <https://sphinx-book-theme.readthedocs.io/en/stable/index.html>`_.
+
+All documentation source files are neatly organized in the ``docs/`` directory within the project's repository.
 
 
-Manual documentation creation
------------------------------
+Local documentation generation
+--------------------------------
 
-After editing the source files, the documentation can be generated locally:
+To build the documentation locally, follow the next three steps.
+
+1. Set up the python environment for building the documentation
+
+	The python packages required to build the documentation are listed in the `requirements.txt <https://github.com/ltelab/disdrodb/blob/main/docs/requirements.txt>`_ file.
+
+	For an efficient setup, we recommend creating a dedicated virtual environment.
+	Navigate to the ``docs/`` directory and execute the following command.
+	This will create a new environment and install the required packages and
+	`pandoc <https://pandoc.org/>`_, a versatile document conversion tool:
+
+	.. code-block:: bash
+
+		conda create -f environment.yaml
+
+2. Activate the virtual environment
+
+	Once the environment is ready, activate it using:
+
+	.. code-block:: bash
+
+		conda activate build-doc-disdrodb
+
+3. Generate the documentation
+
+	With the environment set and activated, you're ready to generate the documentation.
+	Execute:
+
+	.. code-block:: bash
+
+		make clean html
+
+	This command will build the HTML version of the documentation.
+	It first cleans previous builds (``make clean``) and then generates fresh documentation (``html``).
+
+	.. note:: It's important to review the output of the command. Look out for warnings or errors and address them to ensure the documentation is accurate and complete.
+
+	By following these steps, you should have a local version of the DISDRODB documentation
+	in the ``docs/build/html/`` directory, ready for review or deployment!
 
 
-.. code-block:: bash
+Automatic documentation deployment
+------------------------------------
 
-	cd docs
-	make html
-
-
-The output of the previous command should be checked for warnings and errors. If the code is changed (new functions or classes), then the DISDRODB documentation files located in ``docs/source/api`` are automatically regenerated when building the documentation using the command above.
-
-
-Automatic (GitHub) documentation creation
-------------------------------------------
-
-
-One webhook is defined in the repository to trigger the publication process to readthedoc.io.
[…]
