+++ pyproject.toml
@@ -38,9 +38,6 @@ dynamic = ["version"]
 
 [project.optional-dependencies]
 dev = [
-	"sphinx",
-	"sphinx_book_theme",
-	"nbsphinx",
 	"jupyter",
 	"pre-commit",
 	"pytest",
