+++ docs/source/l0_processing.rst
@@ -1,3 +1,5 @@
+.. _l0_processing:
+
 ============================
 DISDRODB L0 processing
 ============================
