+++ docs/source/installation.rst
@@ -1,3 +1,5 @@
+.. _installation:
+
 =========================
 Installation
 =========================
@@ -63,6 +65,9 @@ or `conda <https://docs.conda.io/en/latest/>`__ (recommended).
    virtualenv -p python3 disdrodb-pyXXX
    source disdrodb-pyXXX/bin/activate
 
+
+.. _installation_standard:
+
 Installation for standard users
 ==================================
 
@@ -90,26 +95,27 @@ Please install the package in the virtual environment you created before !
    pip install disdrodb
 
 
+.. _installation_contributor:
+
 Installation for contributors
 ================================
 
-The latest disdrodb version is available on the GitHub repository `disdrodb <https://github.com/ltelab/disdrodb>`_.
+The latest disdrodb version is available on the GitHub repository `disdrodb <https://github.com/ltelab/disdrodb>`__.
 You can install the package in editable mode, so that you can modify the code and see the changes immediately.
 Here below we provide the steps to install the package in editable mode.
 
 Clone the repository from GitHub
 ......................................
 
-According to the `contributors guidelines <https://disdrodb.readthedocs.io/en/latest/contributors_guidelines.html>`__,
-you should first
-`create a fork into your personal GitHub account <https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/working-with-forks/fork-a-repo>__`.
+According to the :ref:`contributors guidelines <contributor_guidelines>`, you should first
+`create a fork into your personal GitHub account <https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/working-with-forks/fork-a-repo>`__.
 
 Then create a local copy of the repository you forked with:
 
 .. code-block:: bash
 
-    git clone https://github.com/<your-account>/disdrodb.git
-    cd disdrodb
+   git clone https://github.com/<your-account>/disdrodb.git
+   cd disdrodb
 
 Create the development environment
 ......................................
@@ -137,7 +143,7 @@ Install the disdrodb package in editable mode by executing the following command
 
 .. code-block:: bash
 
-	pip install -e .
+	pip install -e ".[dev]
 
 
 Install pre-commit code quality checks
@@ -147,23 +153,17 @@ Install the pre-commit hook by executing the following command in the disdrodb r
 
 .. code-block:: bash
 
-   pip install pre-commit
    pre-commit install
 
 
-The pre-commit hooks are scripts executed automatically in every commit
-to identify simple code quality issues. When an issue is identified
-(the pre-commit script exits with non-zero status), the hook aborts the
-commit and prints the error. Currently, DISDRODB tests that the
-code to be committed complies with `black's  <https://github.com/psf/black>`__ format style
-and the `ruff <https://github.com/charliermarsh/ruff>`__ linter.
-
-In case that the commit is aborted, you only need to run `black` and `ruff` through your code.
-This can be done by running ``black .`` and ``ruff check .`` or alternatively with ``pre-commit run --all-files``.
-The latter is recommended since it indicates if the commit contained any formatting errors (that are automatically corrected).
+Pre-commit hooks are automated scripts that run during each commit to detect basic code quality issues.
+If a hook identifies an issue (signified by the pre-commit script exiting with a non-zero status), it halts the commit process and displays the error messages.
 
 .. note::
-	The software version of pre-commit tools is defined into the `.pre-commit-config.yaml` file.
+	The software version of pre-commit hooks is defined into the `.pre-commit-config.yaml <https://github.com/ltelab/disdrodb/blob/main/.pre-commit-config.yaml>`__ file.
+
[…]
