+++ docs/source/contribute_data.rst
@@ -92,7 +92,7 @@ Step 1: Fork and download the DISDRODB Metadata Archive
    `"Create a pull request documentation" <https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-a-pull-request>`__.
 
    If you struggle with this process, do not hesitate to raise an `issue <https://github.com/ltelab/disdrodb-data/issues/new/choose>`__
-   or ask in the `disdrodb slack channel <https://join.slack.com/t/disdrodbworkspace/shared_invite/zt-25l4mvgo7-cfBdXalzlWGd4Pt7H~FqoA>`__ so that we can help !
+   or ask in the `DISDRODB Slack Workspace <https://join.slack.com/t/disdrodbworkspace/shared_invite/zt-25l4mvgo7-cfBdXalzlWGd4Pt7H~FqoA>`__ so that we can help !
 
 
 .. _step2:
@@ -139,7 +139,7 @@ to install your local version of the `disdrodb python package  <https://github.c
    `"Create a pull request documentation" <https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-a-pull-request>`__.
 
    If you struggle with this process, do not hesitate to raise an `issue <https://github.com/ltelab/disdrodb/issues/new/choose>`__
-   or ask in the `disdrodb slack channel <https://join.slack.com/t/disdrodbworkspace/shared_invite/zt-25l4mvgo7-cfBdXalzlWGd4Pt7H~FqoA>`__ so that we can help !
+   or ask in the `DISDRODB Slack Workspace <https://join.slack.com/t/disdrodbworkspace/shared_invite/zt-25l4mvgo7-cfBdXalzlWGd4Pt7H~FqoA>`__ so that we can help !
 
 
 7. Finally, install the disdrodb python package in editable mode using:
@@ -201,7 +201,7 @@ of the stations you wish to contribute.
 The metadata YAML file contains information of the station (e.g. type of raw data, type of device, geolocation, ...) which is
 required for the correct processing and integration of the station into the DISDRODB archive.
 
-The list and description of the metadata keys is available in the `Metadata <https://disdrodb.readthedocs.io/en/latest/metadata.html>`_ section.
+The list and description of the metadata keys is available in the :ref:`Metadata <metadata>` section.
 
 There are 7 metadata keys for which it is mandatory to specify the value:
 
@@ -210,7 +210,7 @@ There are 7 metadata keys for which it is mandatory to specify the value:
 * the ``station_name`` must be the same as the name of the metadata YAML file without the .yml extension
 * the ``sensor_name`` must be one of the implemented sensor configurations. See ``disdrodb.available_sensor_names()``.
   If the sensor which produced your data is not within the available sensors, you first need to add the sensor
-  configurations. For this task, read the section `Add new sensor configs <https://disdrodb.readthedocs.io/en/latest/sensor_configs.html>`_
+  configurations. For this task, read the section :ref:`Add new sensor configs <sensor_configurations>`
 * the ``raw_data_format`` must be either ``'txt'`` or ``'netcdf'``. ``'txt'`` if the source data are text/ASCII files. ``'netcdf'`` if source data are netCDFs.
 * the ``platform_type`` must be either ``'fixed'`` or ``'mobile'``. If ``'mobile'``, the DISDRODB L0 processing accepts latitude, longitude and altitude coordinates to vary with time.
 * the ``reader`` name is essential to enable to select the correct reader when processing the station.
@@ -329,7 +329,7 @@ Step 7: Implement the reader
 Once the DISDRODB directory structure, the raw data and the metadata are set up, you are ready to implement the DISDRODB reader of your data.
 
[…]

[…]
