+++ CONTRIBUTING.rst
@@ -35,7 +35,7 @@ Here are some guidelines to facilitate this process:
 
 1. Start with a conversation
 
-   Before start coding, open a `GitHub discussion <https://github.com/ltelab/disdrodb/discussions>`__, a `GitHub Issue <https://github.com/ltelab/disdrodb/issues/new/choose>`__ or
+   Before start coding, open a `GitHub Discussion <https://github.com/ltelab/disdrodb/discussions>`__, a `GitHub Feature Request Issue <https://github.com/ltelab/disdrodb/issues/new/choose>`__ or
    just start a discussion in the `DISDRODB Slack Workspace <https://join.slack.com/t/disdrodbworkspace/shared_invite/zt-25l4mvgo7-cfBdXalzlWGd4Pt7H~FqoA>`__.
    These channels of communication provides an opportunity to gather feedback, understand the project's current state, and improve your contributions.
 
@@ -50,26 +50,21 @@ Here are some guidelines to facilitate this process:
 
 By following these steps, you not only enhance the quality and relevance of your contribution but also become an integral part of the project's collaborative ecosystem.
 
-If you have any questions, please do not hesitate to ask in the `GitHub discussions <https://github.com/ltelab/disdrodb/discussions>`__ or in the
-`Slack workspace <https://join.slack.com/t/disdrodbworkspace/shared_invite/zt-25l4mvgo7-cfBdXalzlWGd4Pt7H~FqoA>`__.
+If you have any questions, please do not hesitate to ask in the `GitHub Discussions <https://github.com/ltelab/disdrodb/discussions>`__ or in the
+`DISDRODB Slack Workspace <https://join.slack.com/t/disdrodbworkspace/shared_invite/zt-25l4mvgo7-cfBdXalzlWGd4Pt7H~FqoA>`__.
 
 
 Issue Reporting Guidelines
 -----------------------------
 
-For a structured and efficient issue reporting process, please adhere to the following guidelines:
+To facilitate and enhance the issue reporting process, it is important to utilize the predefined GitHub Issue Templates.
+These templates are designed to ensure you provide all the essential information in your report, allowing for a faster and more effective response from the maintainers.
+You can access and use these templates by visiting the `GitHub Issue Templates page here <https://github.com/ltelab/disdrodb/issues/new/choose>`__.
 
-1. Utilize GitHub Issue Templates
-
-Make use of the predefined GitHub Issue Templates available in the repository to report an issue.
-This ensures that all necessary details are provided, facilitating quicker and more effective responses.
-Access the `GitHub Issue Templates here <https://github.com/ltelab/disdrodb/issues/new/choose>`__.
-
-2. Requesting New Templates
-
-If none of the available templates seem appropriate for the issue you are facing, do not hesitate to request a new template.
-This helps us improve and cater to a wider range of issues. Simply open a general issue and specify the need for a new template,
-providing details about what the new template should cover.
+However, if you find that the existing templates don't quite match the specifics of the issue you're encountering, please feel free to suggest a new template.
+Your feedback is invaluable in refining our processes and ensuring we address a broader spectrum of concerns.
+To do this, simply create a general issue in the repository, clearly stating that you're requesting a new template and include detailed suggestions about what this new template should entail.
+This proactive approach helps us continuously evolve and better serve the needs of the project and its contributors.
 
 
 Contributing process
@@ -99,6 +94,7 @@ to create the local copy of the disdrodb repository, set up the developing envir
 Once you have have a local copy of the disdrodb repository on your machine, you are ready to
 contribute to the project!
 
+
 2. Create a new branch
 ~~~~~~~~~~~~~~~~~~~~~~~
 
[…]
