+++ .github/workflows/tests.yml
@@ -46,7 +46,7 @@ jobs:
           pytest # See pyproject.toml for cov options
 
       - name: Upload coverage reports to Codecov
-        uses: codecov/codecov-action@v3
+        uses: codecov/codecov-action@v4
         if: matrix.coverage
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
