+++ .github/workflows/tests.yml
@@ -26,7 +26,7 @@ jobs:
             coverage: true
 
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
 
       - name: Set up micromamba and install package dependencies
         uses: mamba-org/setup-micromamba@v1
