+++ docs/source/contribute_data.rst
@@ -37,17 +37,17 @@ The name you adopt for the ``<DATA_SOURCE>`` and ``<CAMPAIGN_NAME>`` will be use
 
 Here below we provide a detailed description of the steps to follow to contribute your data to DISDRODB:
 
-* `Step 1 <#step1>`_: Fork and download the DISDRODB Metadata Archive
-* `Step 2 <#step2>`_: Define the local DISDRODB Data Archive base directory
-* `Step 3 <#step3>`_: Add metadata in the local DISDRODB Data Archive
-* `Step 4 <#step4>`_: Add the raw data in the local DISDRODB Data Archive
-* `Step 5 <#step5>`_: Fork and install the disdrodb python package in editable mode
-* `Step 6 <#step6>`_: Define the reader name and add a prototype reader to the disdrodb python package
-* `Step 7 <#step7>`_: Implement the reader for your data
-* `Step 8 <#step8>`_: Test the reader by running the DISDRODB L0 processing
-* `Step 9 <#step9>`_: Add reader testing files to the disdrodb python package
-* `Step 10 <#step10>`_: Upload your raw data on Zenodo and link it to the DISDRODB Decentralized Data Archive
-* `Step 11 <#step11>`_: Test the download and DISDRODB L0 processing of the stations you just contributed
+* :ref:`Step 1 <step1>`: Fork and download the DISDRODB Metadata Archive
+* :ref:`Step 2 <step2>`: Define the local DISDRODB Data Archive base directory
+* :ref:`Step 3 <step3>`: Add metadata in the local DISDRODB Data Archive
+* :ref:`Step 4 <step4>`: Add the raw data in the local DISDRODB Data Archive
+* :ref:`Step 5 <step5>`: Fork and install the disdrodb python package in editable mode
+* :ref:`Step 6 <step6>`: Define the reader name and add a prototype reader to the disdrodb python package
+* :ref:`Step 7 <step7>`: Implement the reader for your data
+* :ref:`Step 8 <step8>`: Test the reader by running the DISDRODB L0 processing
+* :ref:`Step 9 <step9>`: Add reader testing files to the disdrodb python package
+* :ref:`Step 10 <step10>`: Upload your raw data on Zenodo and link it to the DISDRODB Decentralized Data Archive
+* :ref:`Step 11 <step11>`: Test the download and DISDRODB L0 processing of the stations you just contributed
 
 Before going down the road, please also have a look at the `Contributors Guidelines <contributors_guidelines.html>`_.
 
@@ -218,7 +218,7 @@ There are 7 metadata keys for which it is mandatory to specify the value:
 Please take care of the following points when editing the metadata files:
 
 *  Do not eliminate metadata keys for which no information is available !
-*  You will define the ``reader`` name in `Step 6 <#step6>`_ along with the implementation of the reader
+*  You will define the ``reader`` name in :ref:`Step 6 <step6>` along with the implementation of the reader
 *  The station metadata YAML file must keep the name of the station (i.e. ``<station_name>.yml``)
 *  For each ``<station_name>`` directory in the ``/data`` directory there must be an equally named ``<station_name>.yml`` file in the ``/metadata`` directory.
 
@@ -342,7 +342,7 @@ which will guide you to the definition of the 4 relevant DISDRODB reader compone
 
 * The ``df_sanitizer_fun()`` function defining the processing to apply on the read dataframe in order for the dataframe to match the DISDRODB standards. The dataframe which is returned by the ``df_sanitizer_fun`` must have only columns compliant with the DISDRODB standards !
 
-When this 4 components are correctly defined, they can be transcribed into the reader you defined in `Step 6 <#step6>`_ and you are ready
[…]
