+++ docs/environment.yaml
@@ -3,7 +3,13 @@ channels:
   - conda-forge
 dependencies:
   - python>=3.9
+  - disdrodb
+  - jupyter
+  - pydantic>=2.0.0
+  - pre-commit
   - pandoc
-  - pip
-  - pip:
-      - -r requirements.txt
+  - nbsphinx==0.9.3
+  - sphinx==5.1.1
+  - sphinx-book-theme==1.1.0
+  - sphinx-mdinclude==0.5.3
+  - sphinx-gallery
