+++ readthedocs.yml
@@ -1,20 +1,0 @@
-# Read the Docs configuration file
-# See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
-version: 2
-# Build documentation in the docs/ directory with Sphinx
-build:
-  os: ubuntu-20.04
-  tools:
-    python: "3.10"
-
-
-sphinx:
-  configuration: docs/source/conf.py
-
-# Optionally build your docs in additional formats such as PDF and ePub
-formats:
-  - pdf
-
-python:
-   install:
-   - requirements: docs/requirements.txt
