+++ .readthedocs.yaml
@@ -1,0 +1,28 @@
+# Read the Docs configuration file
+# See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
+version: 2
+# Build documentation in the docs/ directory with Sphinx
+build:
+  os: ubuntu-22.04
+  tools:
+    python: "mambaforge-22.9"
+
+
+sphinx:
+  configuration: docs/source/conf.py
+
+# Optionally build your docs in additional formats such as PDF and ePub
+formats:
+  - pdf
+
+# Optional but recommended, declare the Python requirements required
+# to build your documentation
+# See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
+
+# python:
+#    install:
+#    - requirements: docs/requirements.txt
+
+# See https://docs.readthedocs.io/en/stable/guides/conda.html#making-builds-faster-with-mamba
+conda:
+  environment: docs/environment.yaml
