### 2024-02-06 #156 @lwasser
hi @ghiggi 👋 I wanted to welcome you to pyOpenSci! We have seen this submission and an editor will get back to you with some initial checks soon. In the meantime if you have any questions you can ask here or in our [discourse](https://pyopensci.discourse.group/).
