### 2024-02-07 #156 @isabelizimm
Hello and welcome to pyOpenSci!!!

---
---

Few small nits as I am going through the repository. These nits are non-blocking from continuing on with your review, but thought I would call them out.

- When viewing on GitHub, all :ref: links in [CONTRIBUTING.rst](https://github.com/ltelab/disdrodb/blob/main/CONTRIBUTING.rst#id6) are not clickable. Additionally, the link to https://github.com/ltelab/disdrodb/blob/main/static/vs_code_settings.png is broken in `3. Work on your changes` -> `Code Documentation`. They seem to work fine in the website, but IMO documents are often ingested in the GitHub UI.
- There are unnecessary leading spaces in the readme code chunks.
- https://disdrodb.readthedocs.io/en/latest/installation.html#install-the-disdrodb-package-in-editable-mode is missing a closing quotation mark.

An overall impression I've received from going through your packages is how much effort you all have put into documenting the tricky steps that people might not be super familiar with. This is not trivial work, so I just wanted to call out how impressed I was!

The one piece that I would like to open up for discussion is

> The package imports properly into a standard Python environment `import package`.

I do agree that `disdrodb` itself imports easily from PyPI. However, it does not seem like the PyPI package itself is useable without cloning https://github.com/ltelab/disdrodb-data, which is _not_ able to be imported via PyPI. Are you able to give some more context on why this was a different repository vs. generated from the package? It does feel a bit awkward currently, but I'm sure there were tradeoffs that impacted the decision of having this structure so I'm curious to hear why this infrastructure was chosen!
