### 2024-02-08 #156 @ghiggi
Hi @isabelizimm,

Thanks a lot for going through the software and thoroughly reviewing the documentation !
Here below I answer to the points you raised.

**Documentation Glitches**

I've addressed the glitches you encountered in the `CONTRIBUTING.rst `and the `README.md`. For the `CONTRIBUTING.rst`, I modified the references with URLs of the online documentation to allow compilation by `Sphinx` while ensuring the file remains readable on GitHub. As an alternative, we could considering moving `CONTRIBUTING.rst `to the `docs/source `directory and adding a `symlink` of the `CONTRIBUTING` file to the base of the repository. What is your opinion?

**Reasons Behind the Creation of the disdrodb-data Repository**

The decision to separate the `disdrodb-data` repository was made taking into consideration the project's structure and the future development.

The `disdrodb-data` repository, or the DISDRODB Metadata Archive, serves as an online platform to track DISDRODB stations, with each station described by a metadata YAML file. This file includes a standard set of metadata fields detailing the instrument specifics, location, and the URL to the online data storage repository.

By cloning the `disdrodb-data` repository to a local machine, users essentially create a storage space for DISDRODB data downloads. This setup allows for easy updates to station metadata or the addition of new stations with a simple `git pull`, ensuring the local data remains up-to-date and accurate. The community's ability to report specific timestamps or periods of sensor malfunctions in the DISDRODB Metadata Archive also enable for recursive improvement of data quality, while maintaining the DISDRODB product chain's transparency and reproducibility.

Given the potential high-frequency of updates to the `disdrodb-data` metadata repository compared to the less frequent updates of the `disdrodb` python package routines, separating these components was a spontaneous decision. This separation also considers GitHub's file size limits and the anticipated expansion of our archive.

It's worth noting that the `disdrodb` python package will, in the future, host processing algorithms for the DSD that do not necessarily require the `disdrodb-data`. This flexibility allows for broader application and utility of the software independent of the metadata repository.

We believe this structure best supports the project's current needs and future growth, balancing ease of use with the complexity of data management and software development.

Thank you once again for your constructive feedback and for opening up this discussion.

I am looking forward for additional questions and suggestions you might have :)
