### 2024-02-21 #156 @isabelizimm
Hey there-- I've been thinking about this a bunch, and also have [solicited some advice from the awesome pyOpenSci community](https://pyopensci.slack.com/archives/C020E991HU6/p1707920427069219) on their experiences with storing data outside a package! Feel free to chime in on that thread as well if you would like (I couldn't find you on Slack to @)!

From your experience, it seems like the two repo solution has come from:

1. having data that is too big to live inside the package in a reasonable way
2. having data that has a much higher velocity than the package itself
3. users having a comfortability with tools (or at least providing documentation to help users with the sticking points)

I do think that this setup feels a bit awkward, and the complexity might turn people away from the package. BUT I think these problems are super solvable and there's some great tools out there to help!

A few options that might make sense:

- adding functions in `disdrodb` that grab the data from the `disdrodb-data`
- making `disdrodb-data` into its own package, and then `disdrodb` depending on that

### Option 1: adding functions in `disdrodb` that grab the data from the `disdrodb-data`

This is pretty classic "faking data in a package" move. The most common way to import data is something like `from package_name.data import dataset_name`, but it might make sense to implement a sort of `load_dataset('dataset_name')` functionality, similar to HuggingFace's [datasets](https://pypi.org/project/datasets/), which is a front end to fetch data from their hub. That way, maybe you could even have other arguments for date, or something (just brainstorming here 🧠 🎉)!

There is more information and options in a section that is getting added to pyOpenSci's packaging guide https://github.com/pyOpenSci/python-package-guide/pull/110/files (it's currently in draft form, but contains lots of great info).

Tool ideas for loading data:

- [pooch](https://github.com/fatiando/pooch): grab data from any url! You could even host the data in zenodo, dropbox, or keep it right on GitHub. This one is probably lowest lift, as its prime focus is fetching data from a URL.
- [pins](https://rstudio.github.io/pins-python/): able to read "pins" (where a pin == your data) from "boards" such as a GitHub repo, DropBox, etc. There are less locations, but it does offer metadata on each pin that you can customize with information that you're currently hosting in `.yml` files (disclaimer: I'm also a maintainer 😄)
- [DVC](https://dvc.org/): data version control. This is a bigger, more robust system, but it is especially made for versioning/storing/retrieving data.

### Option 2: making `disdrodb-data` into its own package, and then `disdrodb` depending on that

This is another classic move for data-- a good example of this is [vega-altair](https://github.com/altair-viz/altair), which has an optional dependency on [vega-datasets](https://pypi.org/project/vega-datasets/) which includes files in it, but also fetches datasets. There are definitely options out there if you don't want to host all the data on GitHub and are concerned about size limits; this solution might end up implementing a little bit of option 1 anyway, if you so desire.

---

Overall, the data can definitely stay in its current repository to help manage the velocity/size/overall organization of files, but I would strongly encourage taking one of these routes to allow users easier access to the data. I understand that it could be a big lift to make this change, but I do think it would be a great investment in the usability of `disdrodb`. Do you have any initial thoughts/feelings/questions on either or both of these options?
