### 2024-02-21 #156 @ghiggi
Hey Isabel,

I tried joining the `pyOpenSci` Slack Workspace but I've been unable because of some access restrictions: I need an invitation I guess 😉 Could you sort that out? You could send the invitation to gionata.ghiggi@epfl.ch. Then we could also chat smoother if something is not clear.

Regarding your message, I noticed there might be some confusion around the `disdrodb-data` repository.
To clarify, it primarily serves as a catalog for metadata, not the actual meteorological measurements !

Maybe we could rename the `disdrodb-data` repository as `disdrodb-metadata` to avoid misinterpretation.
Anyway, to ease the discussion, here below I will refer to such repository as `disdrodb-metadata`.

The actual station data are stored in cloud repositories such Zenodo ([see here for some example](https://zenodo.org/communities/disdrodb?q=&l=list&p=1&s=10)).
Using the URLs information included in the station metadata YAML files of the `disdrodb-metadata` repository, the `disdrodb `package download the stations from the cloud (i.e. Zenodo) using `pooch` 😉

We chose GitHub for the metadata repository for several reasons:
-	It allows for direct, online modifications to the metadata without needing to clone or download/upload files.
-	It provides a robust mechanism for tracking changes over time.
-	It enables us to run specialized CI checks to ensure metadata modifications meet our standards ([see for example this dummy PR](https://github.com/ltelab/disdrodb-data/pull/11)).

When starting the project, we thought about using `DVC` for managing the metadata archive, but decided against it due to the added complexity, which might deter individuals with limited IT experience. Additionally, `DVC` seemed excessive for our needs.

Opting for the two-repo solution was a deliberate choice to facilitate frequent updates to the metadata without necessitating constant releases of the disdrodb Python package each time a metadata file is edited or a new station is added to the metadata archive.

This approach might look unusual, but as far as I know, it’s also the first attempt in the landscape of meteorological (and non?) data management to create a decentralized online data archive infrastructure backed by a GitHub-powered metadata repository 😊

Regarding the options you proposed, they seem to stem from a misunderstanding of the `disdrodb-metadata` repository's role, but I try to answer again to both points:

**Option 2: making disdrodb-metadata into its own package, and then disdrodb depending on that**

The suggestion to package `disdrodb-metadata` seems to misunderstand its purpose.

 If I understand the suggestion correctly, you assumed `disdrodb-data` contained the actual meteorological data, right?

The `disdrodb-metadata` repository does not contain any code and any meteorological measurement.
The` disdrodb-metadata` repository holds stations metadata files in a structured directory. By cloning this repository, we want the user to have such structure on disk so that the `disdrodb `package can exploit the metadata files (and the specified stations URLs) to download the actual stations data which are instead stored in online scientific data repositories.

We want this (cloned) directory to be kept in synchronization with the main repository so that any repo update can be immediately be available to the user, without requiring new package releases. The packaging solution goes in contrast with the potential high frequency of repo updates.

**Option 1: adding functions in disdrodb that grab the data from the disdrodb-metadata**

The `disdrodb-metadata` repository contains YAML metadata files that list and describe weather stations, including URLs for actual data stored online, not the data measurements themselves.

Users interested in downloading disdrometer data simply:

1. Install `disdrodb` with: `pip install disdrodb` or `conda install disdrodb`
2. Clone `disdrodb-metadata` via a terminal command: `git clone git@github.com:ltelab/disdrodb-metadata.git`
3. Indicate the local path to the `/disdrodb-metadata/DISDRODB` directory in `DISDRODB_BASE_DIR`.
4. The `disdrodb `software then fetches and downloads data from online sources like Zenodo into the cloned local `disdrodb-metadata` directory structure (that’s why we have called the repo `disdrodb-data` instead of `disdrodb-metadata`: the repo name also determines the name of the directory where then the data archive is downloaded locally)
5. Updates, including new stations, are easily integrated with a simple `git pull`, minimizing the need for deep knowledge of Git beyond initial setup.

I recognize that using (and installing) `git `could introduce some complexity, but we only require the user to type one `git `command in the terminal, nothing more. Knowledge of the git workflow is not required.

To eliminate the need for `git` for the average user, I see the following options:

1.	We could suggest to download a zip file of the repository (i.e. with this link `https://github.com/ltelab/disdrodb-data/archive/refs/heads/main.zip`), then unzipping and moving it to the desired location of the local DISDRODB archive (defined by `DISDRODB_BASE_DIR`).
Alternatively, we could integrate a utility function in the `disdrodb` Python package to download and unzip the `disdrodb-metadata` repository directly to the specified `DISDRODB_BASE_DIR` on the user's local machine.
However, both options would inhibit the metadata archive's ability to update, thus preventing the download of new stations. Additionally, this approach risks overwriting and losing data if the `DISDRODB_BASE_DIR` subdirectories are modified by the user.

2.	Another option to avoid cloning the disdrodb-metadata archive is to enhance the disdrodb Python package with utilities to:
    - Create the necessary directory structure,
    - Connect to and download/update metadata from the GitHub` disdrodb-metadata` repository upon user request, albeit with the risk of overwriting any local changes,

   Implementing this would demand significant effort for functionality that git already offers efficiently and reliably.

   Moreover, it could complicate in several ways the process for contributors wanting to add new stations to the DISDRODB
   data archive.  For example, if we adopt option 2, a new data contributor should prepare the metadata in the disdrodb-
   created directories (which are not synchronized with disdrodb-metadata and does not have version control) and then copy- paste it into the dedicated directory which is the branch of the forked repository of the disdrodb-metadata. If there are edits to do, it must copy back stuffs to the other directory, check they works, then copy back, etc ...

----------------------------------------------------------------------------------------------------------------------------------------

I hope you don't feel overwhelmed by all that I've just written down :sweat_smile:

For me, all this discussion it's also a great way to understand how we can further improve the documentation of our software ;)
