### 2024-02-21 #156 @isabelizimm
Oh! It looks like I have misunderstood the `disdrodb-data` repository. I won't have the bandwidth today to go through all the info you've given here, but will get back to you shortly. BUT--I just sent you an invite (I think!) to the Slack since it was a quick task 😄 please do let me know if you did not receive it!
