### 2024-03-01 #156 @isabelizimm
Okay, I am back! Thank you so much for your detailed response, that was super helpful to re-orient my brain to understand this set up.

My opinion is still that the workflow for new users is most ergonomic if they can `pip install` all the necessary artifacts, especially from the close-knit dependency between these two pieces of software. I do think that having `git` commands (even though I know it is just one very small one!) does raise the bar of accessibility to data. I like your suggested fix to fetch the `.zip` file; it seems like a reasonable helper function that should not create an overwhelming amount of work for you.

BUT since the intent of `disdrodb-data` repository is more for metadata for data that might be updated and modified by a user/community, it does make sense that it is maintained separately from the package itself. From the information you provided, I think we are okay to continue on with the review with the setup you have right now. 👍

Let's keep this conversation going with an editor with domain expertise on this specific type of data to help determine what would be most useful to the community, and have reviewers (who will be closer to the codebase) to give an extra eye on usability and "getting started" feelings.

---

So, tldr, no need to make any changes to the setup yet. I am actively looking for an editor, and will let you know when we have someone lined up for you!
