### 2024-03-22 9fadb75
add changelog update command and doc
