### 2024-03-22 2a15749
Changelog update command (#214)
* fix github release ci

* add changelog update command and doc
