+++ CHANGELOG.md
@@ -1,8 +1,6 @@
 # Changelog
 
-All notable changes to **disdrodb** will be documented in this file.
-
-## [0.0.2] - Glorious Green - XXXX-XX-XX
+## Version 0.0.2 - Glorious Green - XXXX-XX-XX
 
 ### Enhancements
 - TODO
@@ -13,6 +11,6 @@ All notable changes to **disdrodb** will be documented in this file.
 ### Bugfixes
 - TODO
 
-## [0.0.1] - DISDRODB Birth Date - 2023-01-30
+## Version 0.0.1 - DISDRODB Birth Date - 2023-01-30
 
 First release of DISDRODB.
