+++ disdrodb/tests/conftest.py
@@ -17,12 +17,19 @@ def create_fake_metadata_directory(base_dir, product, data_source="DATA_SOURCE",
     from disdrodb.api.create_directories import create_metadata_directory
 
     return create_metadata_directory(
-        base_dir=base_dir, product=product, data_source=data_source, campaign_name=campaign_name
+        base_dir=base_dir,
+        product=product,
+        data_source=data_source,
+        campaign_name=campaign_name,
     )
 
 
 def create_fake_station_dir(
-    base_dir, product, data_source="DATA_SOURCE", campaign_name="CAMPAIGN_NAME", station_name="station_name"
+    base_dir,
+    product,
+    data_source="DATA_SOURCE",
+    campaign_name="CAMPAIGN_NAME",
+    station_name="station_name",
 ):
     from disdrodb.api.create_directories import create_station_directory
 
