### 2024-04-02 9aa7c6d
Add flake8-comprehension rules
