### 2024-04-02 5d4e681
Add custom ruff rules
