### 2024-04-02 374ae94
Add custom ruff rules
