+++ disdrodb/l0/scripts/disdrodb_run_l0b.py
@@ -16,6 +16,7 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 # -----------------------------------------------------------------------------.
 import sys
+from typing import Optional
 
 import click
 
@@ -36,16 +37,16 @@ sys.tracebacklimit = 0  # avoid full traceback error if occur
 @click_base_dir_option
 def disdrodb_run_l0b(
     # L0 disdrodb stations options
-    data_sources: str = None,
-    campaign_names: str = None,
-    station_names: str = None,
+    data_sources: Optional[str] = None,
+    campaign_names: Optional[str] = None,
+    station_names: Optional[str] = None,
     # Processing options
     force: bool = False,
     verbose: bool = True,
     parallel: bool = True,
     debugging_mode: bool = False,
     remove_l0a: bool = False,
-    base_dir: str = None,
+    base_dir: Optional[str] = None,
 ):
     """
     Run the L0B processing of DISDRODB stations.
