### 2024-04-02 c2d2fe2
Add pylint rules
