+++ disdrodb/l0/l0a_processing.py
@@ -733,7 +733,6 @@ def write_l0a(
         log_error(logger=logger, msg=msg, verbose=False)
         raise ValueError(msg)
     # -------------------------------------------------------------------------.
-    return None
 
 
 ####---------------------------------------------------------------------------.
