### 2024-04-02 bd70631
Add pylint rules
