+++ pyproject.toml
@@ -163,7 +163,7 @@ select = [
     # flake8-simplify
     # "SIM",
     # Perflint
-    # "PERF",
+    "PERF",
     # pandas-vet
     # "PD",
     # pylint
