+++ pyproject.toml
@@ -157,7 +157,7 @@ select = [
     # flake8-unused-arguments
     # "ARG",
     # flake8-raise
-    # "RSE",
+    "RSE",
     # flake8-pytest-style
     # "PT",
     # flake8-simplify
