### 2024-04-02 a808120
Add flake8-bugbears rules
