### 2024-04-02 8764dd2
Add flake8-bugbears rules
