+++ disdrodb/tests/test_metadata/test_metadata_manipulation.py
@@ -36,6 +36,6 @@ def test_add_missing_metadata_keys():
     metadata = {"data_source": "valid"}
     result = add_missing_metadata_keys(metadata)
     # Assert that the valid key is still there
-    assert "valid" == result["data_source"]
+    assert result["data_source"] == "valid"
     # Assert that the missing keys are added
     assert "campaign_name" in result
