### 2024-04-02 44b20ed
Fix test_io.py test unit
