### 2024-04-02 457715e
Remove unused arguments
