+++ disdrodb/l0/l0a_processing.py
@@ -138,7 +138,7 @@ def _check_df_sanitizer_fun(df_sanitizer_fun):
         raise ValueError("The `df_sanitizer_fun` must have only `df` as input argument!")
 
 
-def _check_not_empty_dataframe(df, verbose=False):
+def _check_not_empty_dataframe(df):
     if len(df.index) == 0:
         msg = " - The file is empty and has been skipped."
         log_error(logger=logger, msg=msg, verbose=False)
@@ -288,7 +288,7 @@ def remove_issue_timesteps(df, issue_dict, verbose=False):
     return df
 
 
-def cast_column_dtypes(df: pd.DataFrame, sensor_name: str, verbose: bool = False) -> pd.DataFrame:
+def cast_column_dtypes(df: pd.DataFrame, sensor_name: str) -> pd.DataFrame:
     """Convert ``'object'`` dataframe columns into DISDRODB L0A dtype standards.
 
     Parameters
@@ -297,8 +297,6 @@ def cast_column_dtypes(df: pd.DataFrame, sensor_name: str, verbose: bool = False
         Input dataframe.
     sensor_name : str
         Name of the sensor.
-    verbose : bool
-        Whether to verbose the processing. The default is ``False``.
 
     Returns
     -------
@@ -328,7 +326,7 @@ def cast_column_dtypes(df: pd.DataFrame, sensor_name: str, verbose: bool = False
     return df
 
 
-def coerce_corrupted_values_to_nan(df: pd.DataFrame, sensor_name: str, verbose: bool = False) -> pd.DataFrame:
+def coerce_corrupted_values_to_nan(df: pd.DataFrame, sensor_name: str) -> pd.DataFrame:
     """Coerce corrupted values in dataframe numeric columns to ``np.nan``.
 
     Parameters
@@ -337,8 +335,6 @@ def coerce_corrupted_values_to_nan(df: pd.DataFrame, sensor_name: str, verbose: 
         Input dataframe.
     sensor_name : str
         Name of the sensor.
-    verbose : bool
-        Whether to verbose the processing. The default is ``False``.
 
     Returns
     -------
@@ -361,7 +357,7 @@ def coerce_corrupted_values_to_nan(df: pd.DataFrame, sensor_name: str, verbose: 
     return df
 
 
-def strip_string_spaces(df: pd.DataFrame, sensor_name: str, verbose: bool = False) -> pd.DataFrame:
+def strip_string_spaces(df: pd.DataFrame, sensor_name: str) -> pd.DataFrame:
     """Strip leading/trailing spaces from dataframe string columns.
 
     Parameters
@@ -370,8 +366,6 @@ def strip_string_spaces(df: pd.DataFrame, sensor_name: str, verbose: bool = Fals
         Input dataframe.
     sensor_name : str
         Name of the sensor.
-    verbose : bool
-        Whether to verbose the processing. The default is ``False``.
 
     Returns
     -------
@@ -392,7 +386,7 @@ def strip_string_spaces(df: pd.DataFrame, sensor_name: str, verbose: bool = Fals
             try:
                 df[column] = df[column].str.strip()
             except AttributeError:
-                msg = f"AttributeError: The column {column} is not a string/object dtype."
+                msg = f"The column {column} is not a string/object dtype."
                 log_error(logger=logger, msg=msg, verbose=False)
                 raise AttributeError(msg)
     return df
@@ -618,7 +612,7 @@ def process_raw_file(
     )
 
     # - Check if file empty
-    _check_not_empty_dataframe(df=df, verbose=verbose)
+    _check_not_empty_dataframe(df=df)
 
     # - Check dataframe column number matches columns_names
[…]
