### 2024-04-02 b1b9668
Remove unused arguments
