### 2024-04-02 c005697
Implement simple pytest-style guide
