### 2024-04-02 f0f3f70
Implement simple pytest-style guide
