### 2024-04-02 2e9506d
Implement simple pytest-style guide
