### 2024-04-02 7331a75
Implement simple pytest-style guide
