+++ disdrodb/tests/conftest.py
@@ -152,7 +152,7 @@ def create_fake_raw_data_file(
 
 
 @pytest.fixture()
-def create_test_config_files(request):
+def create_test_config_files(request):  # noqa PT004
     """Create the specified config files into a temporary "test" directory.
 
     This fixture facilitates the creation of configuration files from provided dictionaries.
