+++ disdrodb/l0/readers/EPFL/SAMOYLOV_2019.py
@@ -111,7 +111,7 @@ def reader(
         df["time"] = pd.to_datetime(df["time"], format="%d-%m-%Y %H:%M:%S", errors="coerce")
 
         # Drop rows with None value
-        df = df[~df["TO_BE_SPLITTED"].isnull()]
+        df = df[~df["TO_BE_SPLITTED"].isna()]
 
         # - Drop rows when  'Error in data reading' in TO_BE_SPLITTED column
         bad_indices = df[df["TO_BE_SPLITTED"].str.contains("Error in data reading!")].index
