+++ disdrodb/l0/readers/reader_template.py
@@ -31,7 +31,8 @@ def reader(
     parallel=False,
     debugging_mode=False,
 ):
-    ####----------------------------------------------------------------------.
+    """Reader."""
+    ##------------------------------------------------------------------------.
     ###########################
     #### CUSTOMIZABLE CODE ####
     ###########################
