+++ disdrodb/l0/readers/NETHERLANDS/DELFT.py
@@ -30,6 +30,7 @@ def reader(
     parallel=False,
     debugging_mode=False,
 ):
+    """Reader."""
     ##------------------------------------------------------------------------.
     #### - Define column names
     column_names = ["time", "epoch_time", "TO_BE_PARSED"]
