+++ disdrodb/l0/readers/NCAR/PECAN_FP3.py
@@ -30,7 +30,8 @@ def reader(
     parallel=False,
     debugging_mode=False,
 ):
-    ####----------------------------------------------------------------------.
+    """Reader."""
+    ##------------------------------------------------------------------------.
     #### - Define column names
     column_names = ["TO_SPLIT"]
 
