+++ disdrodb/l0/readers/AUSTRALIA/MELBOURNE_2007_OTT.py
@@ -30,6 +30,7 @@ def reader(
     parallel=False,
     debugging_mode=False,
 ):
+    """Reader."""
     ##------------------------------------------------------------------------.
     #### - Define column names
     column_names = ["TO_BE_PARSED"]
