+++ disdrodb/issue/checks.py
@@ -35,7 +35,6 @@ def _is_numpy_array_string(arr):
     arr : numpy array
         Numpy array to check.
     """
-
     dtype = arr.dtype.type
     return dtype in (np.str_, np.unicode_)
 
@@ -148,7 +147,6 @@ def check_timesteps(timesteps):
 
 def _check_time_period_nested_list_format(time_periods):
     """Check that the time_periods is a list of list of length 2."""
-
     if not isinstance(time_periods, list):
         msg = "'time_periods' must be a list'"
         log_error(logger, msg=msg, verbose=False)
