+++ disdrodb/l0/readers/BRAZIL/GOAMAZON_OTT.py
@@ -30,6 +30,7 @@ def reader(
     parallel=False,
     debugging_mode=False,
 ):
+    """Reader."""
     ##------------------------------------------------------------------------.
     #### - Define column names
     column_names = ["TO_PARSE"]
