+++ disdrodb/l0/scripts/disdrodb_run_l0b.py
@@ -15,6 +15,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 # -----------------------------------------------------------------------------.
+"""Script to run the DISDRODB L0B processing."""
 import sys
 from typing import Optional
 
@@ -57,7 +58,6 @@ def disdrodb_run_l0b(
 
     Parameters
     ----------
-
     data_sources : str
         Name of data source(s) to process.
         The name(s) must be UPPER CASE.
