+++ disdrodb/l0/readers/MEXICO/OH_IIUNAM_nc.py
@@ -31,6 +31,8 @@ def reader(
     parallel=False,
     debugging_mode=False,
 ):
+    """Reader."""
+    ##------------------------------------------------------------------------.
     # Define dictionary mapping dataset variables to select and rename
     dict_names = {
         ### Dimensions
