+++ disdrodb/l0/readers/UK/DIVEN.py
@@ -30,6 +30,8 @@ def reader(
     parallel=False,
     debugging_mode=False,
 ):
+    """Reader."""
+    ##------------------------------------------------------------------------.
     # Define dictionary mapping dataset variables to select and rename
     dict_names = {
         ### Dimensions
