### 2024-04-05 a4adb18
Enable filter available stations by name
