+++ disdrodb/metadata/search.py
@@ -178,58 +178,24 @@ def _get_list_metadata_with_data(base_dir, data_sources=None, campaign_names=Non
         product="RAW",
         data_sources=data_sources,
         campaign_names=campaign_names,
+        station_names=station_names,
     )
 
-    # If no stations available, raise an error
-    if len(list_info) == 0:
-        raise ValueError("No stations are available !")
-
-    if isinstance(station_names, str):
-        station_names = [station_names]
-
-    ### This solution is slow !
-    # if isinstance(station_names, type(None)):
-    #   station_names = [station_name for _, _, station_name in list_info]
-    # metadata_filepaths = [
-    #     define_metadata_filepath(
-    #         product="Raw",
-    #         data_source=data_source,
-    #         campaign_name=campaign_name,
-    #         station_name=station_name,
-    #         base_dir=base_dir,
-    #         check_exists=False,
-    #     )
-    #     for data_source, campaign_name, station_name in list_info if station_name in station_names
-    # ]
-
-    # Get metadata filepaths
-    if isinstance(station_names, list):
-        metadata_filepaths = [
-            define_metadata_filepath(
-                product="Raw",
-                data_source=data_source,
-                campaign_name=campaign_name,
-                station_name=station_name,
-                base_dir=base_dir,
-                check_exists=False,
-            )
-            for data_source, campaign_name, station_name in list_info
-            if station_name in station_names
-        ]
-    else:
-        metadata_filepaths = [
-            define_metadata_filepath(
-                product="Raw",
-                data_source=data_source,
-                campaign_name=campaign_name,
-                station_name=station_name,
-                base_dir=base_dir,
-                check_exists=False,
-            )
-            for data_source, campaign_name, station_name in list_info
-        ]
+    # Retrieve metadata filepaths
+    metadata_filepaths = [
+        define_metadata_filepath(
+            product="RAW",
+            data_source=data_source,
+            campaign_name=campaign_name,
+            station_name=station_name,
+            base_dir=base_dir,
+            check_exists=False,
+        )
+        for data_source, campaign_name, station_name in list_info
+    ]
 
     # If no stations available, raise an error
     if len(metadata_filepaths) == 0:
         raise ValueError("No stations are available !")
+
     return sorted(metadata_filepaths)
