+++ .github/workflows/lint.yml
@@ -14,18 +14,18 @@ jobs:
 
       - name: Ruff
         if: success() || failure()
-        uses: pre-commit/action@v3.0.0
+        uses: pre-commit/action@v3.0.1
         with:
           extra_args: ruff --all-files
 
       - name: Black
         if: success() || failure()
-        uses: pre-commit/action@v3.0.0
+        uses: pre-commit/action@v3.0.1
         with:
           extra_args: black --all-files --verbose
 
       - name: Blackdoc
         if: success() || failure()
-        uses: pre-commit/action@v3.0.0
+        uses: pre-commit/action@v3.0.1
         with:
           extra_args: blackdoc --all-files
