+++ CONTRIBUTING.rst
@@ -237,6 +237,8 @@ Refer to the table below for a comprehensive summary of all CI tools employed to
 +----------------------------------------------------+-----------------------------------------------------------------------------------------------------------------------------------+
 | `CodeFactor <https://www.codefactor.io/>`__        | Automated code review and analysis tools                                                                                          |
 +----------------------------------------------------+-----------------------------------------------------------------------------------------------------------------------------------+
+| `Codacy <https://www.codacy.com/>`__               | Automated code review and analysis tools                                                                                          |
++----------------------------------------------------+-----------------------------------------------------------------------------------------------------------------------------------+
 
 
 5. Code testing with pytest
