### 2024-06-05 6841233
Update precommit rules and docs
