+++ docs/source/index.rst
@@ -1,7 +1,3 @@
-.. d documentation master file, created by
-   sphinx-quickstart on Wed Jul 13 14:44:07 2022.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
 
 Welcome to DISDRODB !
 ======================
