+++ docs/source/installation.rst
@@ -34,13 +34,13 @@ or `conda <https://docs.conda.io/en/latest/>`__ (recommended).
   or `anaconda <https://docs.anaconda.com/anaconda/install/>`__
   if you don't have it already installed.
 
-* Create the `disdrodb-py311` (or any other custom name) conda environment:
+* Create the *disdrodb-py311* (or any other custom name) conda environment:
 
 .. code-block:: bash
 
 	conda create --name disdrodb-py311 python=3.11 --no-default-packages
 
-* Activate the `disdrodb-py311` conda environment:
+* Activate the *disdrodb-py311* conda environment:
 
 .. code-block:: bash
 
@@ -174,10 +174,10 @@ Run DISDRODB on Jupyter Notebooks
 If you want to run disdrodb on a `Jupyter Notebook <https://jupyter.org/>`__,
 you have to take care to set up the IPython kernel environment where disdrodb is installed.
 
-For example, if your conda/virtual environment is named `disdrodb-dev`, run:
+For example, if your conda/virtual environment is named ``disdrodb-dev``, run:
 
 .. code-block:: bash
 
    python -m ipykernel install --user --name=disdrodb-dev
 
-When you will use the Jupyter Notebook, by clicking on `Kernel` and then `Change Kernel`, you will be able to select the `disdrodb-dev` kernel.
+When you will use the Jupyter Notebook, by clicking on ``Kernel`` and then ``Change Kernel``, you will be able to select the ``disdrodb-dev`` kernel.
