+++ docs/environment.yaml
@@ -9,7 +9,8 @@ dependencies:
   - pre-commit
   - pandoc
   - nbsphinx==0.9.3
-  - sphinx==5.1.1
+  - sphinx==7.2.6
   - sphinx-book-theme==1.1.0
   - sphinx-mdinclude==0.5.3
   - sphinx-gallery
+  - docutils
