+++ disdrodb/tests/pytest_files/test_folders_files_creation/mock_file_path.yml
@@ -13,9 +13,9 @@
 # time_period:
 # - ['2018-08-01 12:00:00', '2018-08-01 14:00:00']
 # - ['2018-08-01 15:44:30', '2018-08-01 15:59:31']
-# - ['2018-08-02 12:44:30', '2018-08-02 12:59:31'] 
+# - ['2018-08-02 12:44:30', '2018-08-02 12:59:31']
 
 timesteps:
-- '2018-12-07 14:15:00'
+  - "2018-12-07 14:15:00"
 time_periods:
-- ['2018-08-01 12:00:00', '2018-08-01 14:00:00']
+  - ["2018-08-01 12:00:00", "2018-08-01 14:00:00"]
