+++ disdrodb/tests/pytest_files/test_folders_files_creation/metadata/123.yml
@@ -1,55 +1,55 @@
-data_source: ''
-campaign_name: ''
-station_name: ''
-sensor_name: ''
-reader: ''
+data_source: ""
+campaign_name: ""
+station_name: ""
+sensor_name: ""
+reader: ""
 raw_data_format: txt
 platform_type: fixed
-source: ''
-source_convention: ''
-source_processing_date: ''
-title: ''
-description: ''
-project_name: ''
-keywords: ''
-summary: ''
-history: ''
-comment: ''
-station_id: ''
-location: ''
-country: ''
-continent: ''
+source: ""
+source_convention: ""
+source_processing_date: ""
+title: ""
+description: ""
+project_name: ""
+keywords: ""
+summary: ""
+history: ""
+comment: ""
+station_id: ""
+location: ""
+country: ""
+continent: ""
 latitude: -9999
 longitude: -9999
 altitude: -9999
-deployment_status: ''
-deployment mode: ''
-platform_protection: ''
-platform_orientation: ''
-sensor_long_name: ''
-sensor_manufacturer: ''
-sensor_wavelength: ''
-sensor_serial_number: ''
-firmware_iop: ''
-firmware_dsp: ''
-firmware_version: ''
-sensor_beam_length: ''
-sensor_beam_width: ''
-sensor_nominal_width: ''
-measurement_interval: ''
-calibration_sensitivity: ''
-calibration_certification_date: ''
-calibration_certification_url: ''
-contributors: ''
-authors: ''
-authors_url: ''
-contact: ''
-contact_information: ''
-acknowledgement: ''
-references: ''
-documentation: ''
-website: ''
-institution: ''
-source_repository: ''
-license: ''
-doi: ''
+deployment_status: ""
+deployment mode: ""
+platform_protection: ""
+platform_orientation: ""
+sensor_long_name: ""
+sensor_manufacturer: ""
+sensor_wavelength: ""
+sensor_serial_number: ""
+firmware_iop: ""
+firmware_dsp: ""
+firmware_version: ""
+sensor_beam_length: ""
+sensor_beam_width: ""
+sensor_nominal_width: ""
+measurement_interval: ""
+calibration_sensitivity: ""
+calibration_certification_date: ""
+calibration_certification_url: ""
+contributors: ""
+authors: ""
+authors_url: ""
+contact: ""
+contact_information: ""
+acknowledgement: ""
+references: ""
+documentation: ""
+website: ""
+institution: ""
+source_repository: ""
+license: ""
+doi: ""
