+++ disdrodb/tests/pytest_files/test_folders_files_creation/DISDRODB/Processed/DATA_SOURCE/CAMPAIGN_NAME/metadata/STATIONID.yml
@@ -1,15 +1,15 @@
-title: 'CABAUW'
-description: ''
-source: ''
-history: ''
-conventions: ''
-campaign_name: 'DELFT'
-project_name: ''
-station_id: 'STATIONID'
-station_name: 'Cabauw'
-location: ''
-country: 'Netherlands'
-continent: 'Europe'
+title: "CABAUW"
+description: ""
+source: ""
+history: ""
+conventions: ""
+campaign_name: "DELFT"
+project_name: ""
+station_id: "STATIONID"
+station_name: "Cabauw"
+location: ""
+country: "Netherlands"
+continent: "Europe"
 latitude: 51.96832449
 longitude: 4.92921567
 altitude: 1
@@ -19,25 +19,25 @@ EPSG: 4326
 latitude_unit: DegreesNorth
 longitude_unit: DegreesEast
 altitude_unit: MetersAboveSeaLevel
-sensor_name: 'OTT_Parsivel2'
-sensor_long_name: 'OTT Hydromet Parsivel 2'
-sensor_wavelength: ''
-sensor_serial_number: ''
-firmware_iop: ''
-firmware_dsp: ''
-firmware_version: ''
-sensor_beam_width: ''
-sensor_nominal_width: ''
-measurement_interval: ''
-contributors: ''
-authors: ''
-institution: ''
-references: ''
-documentation: ''
-website: ''
-source_repository: ''
-doi: ''
-contact: ''
-contact_information: ''
-source_data_format: ''
-obs_type: ''
+sensor_name: "OTT_Parsivel2"
+sensor_long_name: "OTT Hydromet Parsivel 2"
+sensor_wavelength: ""
+sensor_serial_number: ""
+firmware_iop: ""
+firmware_dsp: ""
+firmware_version: ""
+sensor_beam_width: ""
+sensor_nominal_width: ""
+measurement_interval: ""
+contributors: ""
+authors: ""
+institution: ""
+references: ""
+documentation: ""
+website: ""
+source_repository: ""
+doi: ""
+contact: ""
+contact_information: ""
+source_data_format: ""
+obs_type: ""
