+++ disdrodb/tests/pytest_files/test_folders_files_creation/DISDRODB/Processed/DATA_SOURCE/CAMPAIGN_NAME/metadata/STATION_NAME.yml
@@ -1,55 +1,55 @@
-data_source: 'DATA_SOURCE'
-campaign_name: 'CAMPAIGN_NAME'
-station_name: 'STATION_NAME'
-sensor_name: 'OTT_Parsivel'
-reader: 'EPFL/EPFL_2009'
+data_source: "DATA_SOURCE"
+campaign_name: "CAMPAIGN_NAME"
+station_name: "STATION_NAME"
+sensor_name: "OTT_Parsivel"
+reader: "EPFL/EPFL_2009"
 raw_data_format: txt
 platform_type: fixed
-source: ''
-source_convention: ''
-source_processing_date: ''
-title: ''
-description: ''
-summary: ''
-history: ''
-keywords: ''
-project_name: ''
-comment: ''
-station_id: ''
-location: ''
-country: ''
-continent: ''
+source: ""
+source_convention: ""
+source_processing_date: ""
+title: ""
+description: ""
+summary: ""
+history: ""
+keywords: ""
+project_name: ""
+comment: ""
+station_id: ""
+location: ""
+country: ""
+continent: ""
 latitude: -9999
 longitude: -9999
 altitude: -9999
-deployment_status: ''
-deployment mode: ''
-platform_protection: ''
-platform_orientation: ''
-sensor_long_name: ''
-sensor_manufacturer: ''
-sensor_wavelength: ''
-sensor_serial_number: ''
-firmware_iop: ''
-firmware_dsp: ''
-firmware_version: ''
-sensor_beam_length: ''
-sensor_beam_width: ''
-sensor_nominal_width: ''
-measurement_interval: ''
-calibration_sensitivity: ''
-calibration_certification_date: ''
-calibration_certification_url: ''
-contributors: ''
-authors: ''
-authors_url: ''
-contact: ''
-contact_information: ''
-acknowledgement: ''
-references: ''
-documentation: ''
-website: ''
-institution: ''
-source_repository: ''
-license: ''
-doi: ''
+deployment_status: ""
+deployment mode: ""
+platform_protection: ""
+platform_orientation: ""
+sensor_long_name: ""
+sensor_manufacturer: ""
+sensor_wavelength: ""
+sensor_serial_number: ""
+firmware_iop: ""
+firmware_dsp: ""
+firmware_version: ""
+sensor_beam_length: ""
+sensor_beam_width: ""
+sensor_nominal_width: ""
+measurement_interval: ""
+calibration_sensitivity: ""
+calibration_certification_date: ""
+calibration_certification_url: ""
+contributors: ""
+authors: ""
+authors_url: ""
+contact: ""
+contact_information: ""
+acknowledgement: ""
+references: ""
+documentation: ""
+website: ""
+institution: ""
+source_repository: ""
+license: ""
+doi: ""
