### 2024-06-05 99ad0c4
Update precommit rules and doc compilation (#217)
* Update precommit rules and docs

* Update docstrings

* Fix dummy typo

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
