### 2024-06-05 d970f7f
Small fixes to docstring
