### 2024-06-27 #156 @ghiggi
Hi @Zeitsperre. I hope you are doing well. Is there any news regarding the review of the software?
I don't want to put pressure but when submitted to pyopensci for review we were planning to have the software manuscript to be published on JOSS by the end of the summer and discuss at a conference in September the next steps :). Do you think is this expectation still realistic?
