### 2024-06-27 #156 @Zeitsperre
Hi @ghiggi,

Apologies for the delay!

I have a few reviewers in mind that will be contacted this week. In terms of the review itself, it can sometimes take more than a month to get all review comments back. Depending on the amount of changes requested, the remaining changes could take some time and effort. I have seen some review turnarounds of a few months at most.

At the very least, for the purposes of your presentation you should be able to use either the Zenodo DOI or perhaps there's a way to park a PyOS DOI ? I'll look into that after contacting the reviewers.
