### 2024-10-01 #156 @lwasser
hey there @Zeitsperre et all. i'm just checking in on this review. Is there anything that i can do to help / support it moving forward. To answer the question above setting this project up with Zenodo is highly recommended.
