+++ .pre-commit-config.yaml
@@ -1,7 +1,7 @@
 ---
 repos:
   - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.5.0
+    rev: v5.0.0
     hooks:
       - id: trailing-whitespace
       - id: end-of-file-fixer
@@ -12,12 +12,12 @@ repos:
       - id: check-ast
       - id: check-added-large-files
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.3.5
+    rev: v0.6.9
     hooks:
       - id: ruff
         args: [--fix]
   - repo: https://github.com/psf/black
-    rev: 24.3.0
+    rev: 24.10.0
     hooks:
       - id: black
         language_version: python3
@@ -27,18 +27,18 @@ repos:
       - id: blackdoc
         additional_dependencies: ["black[jupyter]"]
   - repo: https://github.com/pre-commit/mirrors-prettier
-    rev: "v3.1.0"
+    rev: "v4.0.0-alpha.8"
     hooks:
       - id: prettier
         types_or: [yaml, html, css, scss, javascript, json] # markdown to avoid conflicts with mdformat
   - repo: https://github.com/codespell-project/codespell
-    rev: v2.2.6
+    rev: v2.3.0
     hooks:
       - id: codespell
         types_or: [python, markdown, rst]
         additional_dependencies: [tomli]
   - repo: https://github.com/asottile/pyupgrade
-    rev: v3.15.2
+    rev: v3.18.0
     hooks:
       - id: pyupgrade
   - repo: https://github.com/MarcoGorelli/madforhooks
@@ -47,7 +47,7 @@ repos:
       # - id: conda-env-sorter # conflicts with prettier
       - id: check-execution-order
   - repo: https://github.com/executablebooks/mdformat
-    rev: 0.7.17
+    rev: 0.7.18
     hooks:
       - id: mdformat
         additional_dependencies: [mdformat-gfm, mdformat-black]
@@ -58,7 +58,7 @@ repos:
       - id: nbstripout
         args: [--keep-output]
   - repo: https://github.com/nbQA-dev/nbQA
-    rev: 1.8.5
+    rev: 1.8.7
     hooks:
       - id: nbqa-black
       - id: nbqa-ruff
