+++ disdrodb/l0/scripts/disdrodb_run_l0b_station.py
@@ -20,9 +20,10 @@ from typing import Optional
 
 import click
 
-from disdrodb.l0.routines import click_l0_processing_options, click_remove_l0a_option
+from disdrodb.l0.routines import click_remove_l0a_option
 from disdrodb.utils.scripts import (
     click_base_dir_option,
+    click_processing_options,
     click_station_arguments,
     parse_base_dir,
 )
@@ -35,7 +36,7 @@ sys.tracebacklimit = 0  # avoid full traceback error if occur
 
 @click.command()
 @click_station_arguments
-@click_l0_processing_options
+@click_processing_options
 @click_remove_l0a_option
 @click_base_dir_option
 def disdrodb_run_l0b_station(
