+++ disdrodb/l0/scripts/disdrodb_run_l0a_station.py
@@ -20,9 +20,9 @@ from typing import Optional
 
 import click
 
-from disdrodb.l0.routines import click_l0_processing_options
 from disdrodb.utils.scripts import (
     click_base_dir_option,
+    click_processing_options,
     click_station_arguments,
     parse_base_dir,
 )
@@ -35,7 +35,7 @@ sys.tracebacklimit = 0  # avoid full traceback error if occur
 
 @click.command()
 @click_station_arguments
-@click_l0_processing_options
+@click_processing_options
 @click_base_dir_option
 def disdrodb_run_l0a_station(
     # Station arguments

[…]
