+++ disdrodb/l0/l0_processing.py
@@ -37,11 +37,14 @@ from disdrodb.api.create_directories import (
     create_l0_directory_structure,
 )
 from disdrodb.api.info import infer_path_info_dict
+from disdrodb.api.io import get_filepaths, get_required_product
 from disdrodb.api.path import (
     define_campaign_dir,
+    define_data_dir,
     define_l0a_filepath,
     define_l0b_filepath,
     define_l0b_station_dir,
+    define_l0c_filename,
     define_station_dir,
     get_disdrodb_path,
 )
@@ -53,8 +56,13 @@ from disdrodb.l0.io import (
     read_l0a_dataframe,
 )
 from disdrodb.l0.l0_reader import get_station_reader_function
+from disdrodb.l0.l0b_processing import set_l0b_encodings
+from disdrodb.l0.l0c_processing import (
+    create_daily_file,
+    get_files_per_days,
+)
 from disdrodb.metadata import read_station_metadata
-from disdrodb.utils.decorator import delayed_if_parallel
+from disdrodb.utils.decorator import delayed_if_parallel, single_threaded_if_parallel
 from disdrodb.utils.directories import list_files
 
 # Logger
@@ -66,6 +74,7 @@ from disdrodb.utils.logger import (
     log_info,
     log_warning,
 )
+from disdrodb.utils.writer import write_product
 
 logger = logging.getLogger(__name__)
 
@@ -307,6 +316,84 @@ def _generate_l0b_from_nc(
     return logger_filepath
 
 
+@delayed_if_parallel
+@single_threaded_if_parallel
+def _generate_l0c(
+    day,
+    filepaths,
+    data_dir,
+    processed_dir,  # TODO: REMOVE IN FUTURE
+    campaign_name,
+    station_name,
+    # Processing options
+    force,
+    verbose,
+    parallel,  # this is used only to initialize the correct logger !
+):
+    # -----------------------------------------------------------------.
+    # Define product name
+    product = "L0C"
+
+    # -----------------------------------------------------------------.
+    # Create file logger
+    # TODO: AVOID USE OF processed_dir
+    logger, logger_filepath = create_file_logger(
+        processed_dir=processed_dir,
+        product=product,
+        station_name=station_name,
+        filename=day,
+        parallel=parallel,
+    )
+
+    ##------------------------------------------------------------------------.
+    # Log start processing
+    msg = f"{product} processing for {day} has started."
+    log_info(logger, msg, verbose=verbose)
+
+    ##------------------------------------------------------------------------.
+    ### Core computation
+    try:
+        # Produce L0C dataset
+        ds = create_daily_file(day=day, filepaths=filepaths)
+
+        # Write L0C netCDF4 dataset
+        if ds["time"].size > 1:
+
+            # Get sensor name from dataset
+            sensor_name = ds.attrs.get("sensor_name")
+
+            # Set encodings
+            ds = set_l0b_encodings(ds=ds, sensor_name=sensor_name)
+
+            # Define filepath
+            filename = define_l0c_filename(ds, campaign_name=campaign_name, station_name=station_name)
+            filepath = os.path.join(data_dir, filename)
+
+            # Write to disk
+            write_product(ds, product=product, filepath=filepath, force=force)
+
+        ##--------------------------------------------------------------------.
+        # Clean environment
+        del ds
+
+        # Log end processing
+        msg = f"{product} processing for {day} has ended."
+        log_info(logger, msg, verbose=verbose)
+
+    ##--------------------------------------------------------------------.
+    # Otherwise log the error
+    except Exception as e:
+        error_type = str(type(e).__name__)
+        msg = f"{error_type}: {e}"
+        log_error(logger, msg, verbose=verbose)
+
+    # Close the file logger
[…]

[…]
