+++ disdrodb/l0/l0c_processing.py
@@ -17,48 +17,49 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 # -----------------------------------------------------------------------------.
 """Functions to process DISDRODB L0B files into DISDRODB L0C netCDF files."""
-import logging
-import numpy as np
-import pandas as pd 
-import xarray as xr
 import itertools
-from disdrodb.api.info import get_start_end_time_from_filepaths
+import logging
+
+import numpy as np
+import pandas as pd
+import xarray as xr
 from mydsd.l1.resampling import add_sample_interval
 from mydsd.utils.time import infer_sample_interval, regularize_timesteps
 
+from disdrodb.api.info import get_start_end_time_from_filepaths
 
 logger = logging.getLogger(__name__)
 
 
 def get_files_per_days(filepaths):
     # Retrieve file start_time and end_time
-    files_start_time, files_end_time = get_start_end_time_from_filepaths(filepaths)    
-    
-    # Add tolerance to account for imprecise time logging by the sensors 
-    # - Example: timestep 23:58 that should be 00.00 goes into the next day ... 
+    files_start_time, files_end_time = get_start_end_time_from_filepaths(filepaths)
+
+    # Add tolerance to account for imprecise time logging by the sensors
+    # - Example: timestep 23:58 that should be 00.00 goes into the next day ...
     files_start_time = files_start_time - np.array(60, dtype="m8[s]")
     files_end_time = files_end_time + np.array(60, dtype="m8[s]")
-    
-    # Retrieve file start day and end day 
+
+    # Retrieve file start day and end day
     start_day = files_start_time.min().astype("M8[D]")
     end_day = files_end_time.max().astype("M8[D]") + np.array(1, dtype="m8[D]")
-        
+
     # Create an array with all days in time period covered by the files
-    list_days = np.asanyarray(pd.date_range(start=start_day, end=end_day, freq='D')).astype("M8[D]")
+    list_days = np.asanyarray(pd.date_range(start=start_day, end=end_day, freq="D")).astype("M8[D]")
 
     # Expand dimension to match each day using broadcasting
     files_start_time = files_start_time.astype("M8[D]")[:, np.newaxis]  # shape (n_files, 1)
-    files_end_time = files_end_time.astype("M8[D]")[:, np.newaxis]      # shape (n_files, 1)
-    
+    files_end_time = files_end_time.astype("M8[D]")[:, np.newaxis]  # shape (n_files, 1)
+
     # Create an array of all days
     # - Expand dimension to match each day using broadcasting
     days = list_days[np.newaxis, :]  # shape (1, n_days)
-     
+
     # Use broadcasting to create a boolean matrix indicating which files cover which days
     mask = (files_start_time <= days) & (files_end_time >= days)  # shape (n_files, n_days)
-    
-    # np.sum(mask, axis=0) 
-    
+
+    # np.sum(mask, axis=0)
+
     # Build a mapping from days to file indices
     # For each day (column), find the indices of files (rows) that cover that day
     dict_days = {}
@@ -67,7 +68,7 @@ def get_files_per_days(filepaths):
         file_indices = np.where(mask[:, i])[0]
         if file_indices.size > 0:
             dict_days[str(day)] = filepaths[file_indices].tolist()
-    
+
     return dict_days
 
 
@@ -79,10 +80,10 @@ def find_isel_common_time(da1, da2):
 
 
 def check_same_raw_drop_number_values(list_ds, filepaths):
-    # Retrieve variable to compare 
+    # Retrieve variable to compare
     list_drop_number = [ds["raw_drop_number"].compute() for ds in list_ds]
-    # Compare values 
[…]
