+++ disdrodb/api/path.py
@@ -307,13 +307,227 @@ def define_config_dir(product):
 
 def check_sample_interval(sample_interval):
     if not isinstance(sample_interval, int):
-        raise ValueError("sample_interval must be an integer.")
+        raise ValueError("'sample_interval' must be an integer.")
 
 
 def check_distribution(distribution):
     valid_distributions = ["gamma", "normalized_gamma", "lognormal", "exponential"]
     if distribution not in valid_distributions:
-        raise ValueError(f"Invalid distribution {distribution}. Valid values are {valid_distributions}")
+        raise ValueError(f"Invalid 'distribution' {distribution}. Valid values are {valid_distributions}")
+
+
+def check_rolling(rolling):
+    if not isinstance(rolling, bool):
+        raise ValueError("'rolling' must be a boolean.")
+
+
+def define_product_dir_tree(
+    product,
+    distribution=None,
+    sample_interval=None,
+    rolling=None,
+):
+    """Return the product directory tree.
+
+    Parameters
+    ----------
+    product : str
+        The DISDRODB product. It can be ``"RAW"``, ``"L0A"``, or ``"L0B"``.
+    sample_interval : int, optional
+        The sampling interval in seconds of the product.
+        It must be specified only for product L2E and L2M !
+    rolling : bool, optional
+        Whether the dataset has been resampled by aggregating or rolling.
+        It must be specified only for product L2E and L2M !
+    distribution : str
+        The model of the statistical distribution for the DSD.
+        It must be specified only for product L2M !
+
+    Returns
+    -------
+    data_dir : str
+        Station data directory path
+    """
+    if product.upper() == "RAW":
+        return ""
+    if product.upper() in ["L0A", "L0B", "L0C", "L1"]:
+        return product
+    elif product == "L2E":
+        check_rolling(rolling)
+        check_sample_interval(sample_interval)
+        sample_interval_acronym = get_sample_interval_acronym(seconds=sample_interval, rolling=rolling)
+        return os.path.join(product, sample_interval_acronym)
+    elif product == "L2M":
+        check_rolling(rolling)
+        check_sample_interval(sample_interval)
+        check_distribution(distribution)
+        sample_interval_acronym = get_sample_interval_acronym(seconds=sample_interval, rolling=rolling)
+        distribution_acronym = get_distribution_acronym(distribution)
+        return os.path.join(product, distribution_acronym, sample_interval_acronym)
+    else:
+        raise ValueError("TODO")  # CHECK Product on top !`
+
+
+def define_station_dir_new(
+    product,
+    data_source,
+    campaign_name,
+    station_name,
+    base_dir=None,
+    check_exists=False,
+):  # TODO: IN FUTURE without product --> campaign_dir/station_name/product !
+    """Return the station data directory in the DISDRODB infrastructure.
+
+    Parameters
+    ----------
+    product : str
+        The DISDRODB product. It can be ``"RAW"``, ``"L0A"``, or ``"L0B"``.
+    data_source : str
+        The data source.
+    campaign_name : str
+        The campaign name.
+    station_name : str
+        The station name.
+    base_dir : str, optional
+        The base directory of DISDRODB, expected in the format ``<...>/DISDRODB``.
+        If not specified, the path specified in the DISDRODB active configuration will be used.
+    check_exists : bool, optional
+        Whether to check if the directory exists. By default ``False``.
+
+    Returns
+    -------
+    station_dir : str
+        Station data directory path
+    """
+    base_dir = get_base_dir(base_dir)
+    campaign_dir = get_disdrodb_path(
[…]
