### 2024-12-09 72d444b
Reorganize CLI scripts into single directory
