+++ disdrodb/cli/disdrodb_initialize_station.py
@@ -20,7 +20,7 @@ from typing import Optional
 
 import click
 
-from disdrodb.utils.scripts import click_base_dir_option, click_station_arguments, parse_base_dir
+from disdrodb.utils.cli import click_base_dir_option, click_station_arguments, parse_base_dir
 
 sys.tracebacklimit = 0  # avoid full traceback error if occur
 
