### 2024-12-10 00ff7ee
Refactor for L0C
