+++ disdrodb/scattering/routines.py
@@ -1,0 +1,449 @@
+# -----------------------------------------------------------------------------.
+# Copyright (c) 2021-2023 DISDRODB developers
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+# -----------------------------------------------------------------------------.
+"""Implement PSD scattering routines."""
+
+from itertools import product
+
+import dask
+import numpy as np
+import xarray as xr
+from pytmatrix import orientation, radar, refractive, tmatrix_aux
+from pytmatrix.psd import BinnedPSD, PSDIntegrator
+from pytmatrix.tmatrix import Scatterer
+
+from disdrodb.psd.models import create_psd, get_required_parameters
+from disdrodb.scattering.axis_ratio import check_axis_ratio, get_axis_ratio_method
+
+# Wavelengths for which the refractive index is defined in pytmatrix (in mm)
+wavelength_dict = {
+    "S": tmatrix_aux.wl_S,
+    "C": tmatrix_aux.wl_C,
+    "X": tmatrix_aux.wl_X,
+    "Ku": tmatrix_aux.wl_Ku,
+    "Ka": tmatrix_aux.wl_Ka,
+    "W": tmatrix_aux.wl_W,
+}
+
+
+def available_radar_bands():
+    """Return a list of the available radar bands."""
+    return list(wavelength_dict)
+
+
+def check_radar_band(radar_band):
+    """Check the validity of the specified radar band."""
+    available_bands = available_radar_bands()
+    if radar_band not in available_bands:
+        raise ValueError(f"{radar_band} is an invalid radar band. Valid radar bands: {available_bands}.")
+    return radar_band
+
+
+def get_radar_wavelength(radar_band):
+    """Get the wavelength of a radar band."""
+    wavelength = wavelength_dict[radar_band]
+    return wavelength
+
+
+def initialize_scatterer(wavelength, canting_angle_std=7, D_max=8, axis_ratio="Thurai2007"):
+    """Initialize T-matrix scatterer object for a given wavelength."""
+    # Retrieve custom axis ratio function
+    axis_ratio_func = get_axis_ratio_method(axis_ratio)
+
+    # Retrieve water complex refractive index
+    # - Here we currently assume 10 °C
+    # - m_w_0C and m_w_20C are also available
+    # TODO: should be another dimension ? Or use scatterer.psd_integrator.m_func?
+    water_refractive_index = refractive.m_w_10C[wavelength]
+
+    # ---------------------------------------------------------------.
+    # Initialize Scatterer class
+    scatterer = Scatterer(wavelength=wavelength, m=water_refractive_index)
+    # - Define particle orientation PDF for orientational averaging
+    # --> The standard deviation of the angle with respect to vertical orientation (the canting angle).
+    scatterer.or_pdf = orientation.gaussian_pdf(std=canting_angle_std)
+    # - Define orientation methods
+    # --> Alternatives: orient_averaged_fixed, orient_single
+    scatterer.orient = orientation.orient_averaged_fixed
+
[…]
