+++ disdrodb/l1_env/__init__.py
@@ -14,4 +14,4 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 # -----------------------------------------------------------------------------.
-"""DISDRODB Utils Module."""
+"""Core functions for DISDRODB L1 ENV production."""
