+++ disdrodb/l1/routines.py
@@ -1,0 +1,320 @@
+#!/usr/bin/env python3
+
+# -----------------------------------------------------------------------------.
+# Copyright (c) 2021-2023 DISDRODB developers
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+# -----------------------------------------------------------------------------.
+"""Implement DISDRODB L1 processing."""
+
+import datetime
+import logging
+import os
+import time
+from typing import Optional
+
+import dask
+import xarray as xr
+
+# Directory
+from disdrodb.api.create_directories import (
+    create_product_directory,
+)
+from disdrodb.api.io import get_filepaths, get_required_product
+from disdrodb.api.path import (
+    define_l1_filename,
+    define_logs_dir,
+)
+from disdrodb.configs import get_base_dir
+from disdrodb.utils.decorator import delayed_if_parallel, single_threaded_if_parallel
+
+# Logger
+from disdrodb.utils.logger import (
+    close_logger,
+    create_logger_file,
+    define_summary_log,
+    log_error,
+    log_info,
+)
+from disdrodb.utils.writer import write_product
+
+from disdrodb.l1.processing import generate_l1
+
+logger = logging.getLogger(__name__)
+
+
+def get_l1_options():
+    """Get L1 options."""
+    # - TODO: from YAML
+    # - TODO: as function of sensor name
+
+    # minimum_diameter
+    # --> OTT_Parsivel: 0.2495
+    # --> RD80: 0.313
+    # --> LPM: 0.125 (we currently discard first bin with this setting)
+
+    # maximum_diameter
+    # LPM: 8 mm
+    # RD80: 5.6 mm
+    # OTT: 26 mm
+
+    l1_options = {
+        # Fall velocity option
+        "fall_velocity_method": "Beard1976",
+        # Diameter-Velocity Filtering Options
+        "minimum_diameter": 0.2495,  # OTT Parsivel first two bin no data !
+        "maximum_diameter": 8,
+        "minimum_velocity": 0,
+        "maximum_velocity": 12,
+        "above_velocity_fraction": 0.5,
+        "above_velocity_tolerance": None,
+        "below_velocity_fraction": 0.5,
+        "below_velocity_tolerance": None,
+        "small_diameter_threshold": 1,  # 2
+        "small_velocity_threshold": 2.5,  # 3
+        "maintain_smallest_drops": True,
+    }
+    return l1_options
+
+
+@delayed_if_parallel
+@single_threaded_if_parallel
+def _generate_l1(
+    filepath,
+    data_dir,
+    logs_dir,
+    campaign_name,
+    station_name,
+    # Processing options
+    force,
+    verbose,
+    parallel,  # this is used only to initialize the correct logger !
+):
+    """Generate the L1 product from the DISRODB L0C netCDF file.
+
+    Parameters
+    ----------
+    filepath : str
+        Path to the L0C netCDF file.
+    data_dir : str
+        Directory where the L1 netCDF file will be saved.
+    logs_dir : str
+        Directory where the log file will be saved.
+    campaign_name : str
+        Name of the campaign.
+    station_name : str
+        Name of the station.
+    force : bool
+        If True, overwrite existing files.
+    verbose : bool
+        Whether to verbose the processing.
+
[…]
