### 2024-12-11 deb0972
Add new logger
