+++ disdrodb/l0/readers/BRAZIL/CHUVA_OTT.py
@@ -71,18 +71,18 @@ def reader(
         df = df["TO_PARSE"].str.split(":", expand=True, n=1)
         df.columns = ["ID", "Value"]
 
-        # Drop rows with no values
+        # Select only rows with values
         df = df[df["Value"].astype(bool)]
+        df = df[df["Value"].apply(lambda x: x is not None)]
 
-        # Convert ID to integer
-        # - First convert to numeric and if errors arise (corrupted rows), drop rows
-        df["ID"] = pd.to_numeric(df["ID"], errors="coerce")
-        df = df.dropna(subset="ID")
-        df["ID"] = df["ID"].astype(int)
+        # Drop rows with invalid IDs
+        # - Corrupted rows
+        valid_id_str = np.char.rjust(np.arange(0, 94).astype(str), width=2, fillchar="0")
+        df = df[df["ID"].astype(str).isin(valid_id_str)]
 
         # Create the dataframe with each row corresponding to a timestep
         # - Group rows based on when ID values restart
-        groups = df.groupby((df["ID"].diff() <= 0).cumsum())
+        groups = df.groupby((df["ID"].astype(int).diff() <= 0).cumsum())
 
         # - Reshape the dataframe
         group_dfs = []
