### 2024-12-16 68ca587
Fix issue in L0C processing
