+++ disdrodb/l0/l0c_processing.py
@@ -17,19 +17,23 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 # -----------------------------------------------------------------------------.
 """Functions to process DISDRODB L0B files into DISDRODB L0C netCDF files."""
-import itertools
 import logging
-import shutil
 
 import numpy as np
 import pandas as pd
-import xarray as xr
 
 from disdrodb.api.info import get_start_end_time_from_filepaths
 from disdrodb.l1.resampling import add_sample_interval
-from disdrodb.utils.time import infer_sample_interval, regularize_timesteps
+from disdrodb.utils.logger import log_warning  # , log_info
+from disdrodb.utils.time import (
+    ensure_sorted_by_time,
+    regularize_timesteps,
+)
 
 logger = logging.getLogger(__name__)
+
+
+TOLERANCE_SECONDS = 120
 
 
 def get_files_per_days(filepaths):
@@ -55,9 +59,9 @@ def get_files_per_days(filepaths):
     files_start_time, files_end_time = get_start_end_time_from_filepaths(filepaths)
 
     # Add tolerance to account for imprecise time logging by the sensors
-    # - Example: timestep 23:58 that should be 00.00 goes into the next day ...
-    files_start_time = files_start_time - np.array(60, dtype="m8[s]")
-    files_end_time = files_end_time + np.array(60, dtype="m8[s]")
+    # - Example: timestep 23:59:30 might be 00.00 and goes into the next day file ...
+    files_start_time = files_start_time - np.array(TOLERANCE_SECONDS, dtype="m8[s]")
+    files_end_time = files_end_time + np.array(TOLERANCE_SECONDS, dtype="m8[s]")
 
     # Retrieve file start day and end day
     start_day = files_start_time.min().astype("M8[D]")
@@ -89,74 +93,331 @@ def get_files_per_days(filepaths):
     return dict_days
 
 
-def find_isel_common_time(da1, da2):
+def retrieve_possible_measurement_intervals(metadata):
+    """Retrieve list of possible measurements intervals."""
+    measurement_interval = metadata.get("measurement_interval", [])
+    if isinstance(measurement_interval, (int, float, str)):
+        measurement_interval = [measurement_interval]
+    measurement_intervals = [int(v) for v in measurement_interval]
+    return measurement_intervals
+
+
+def drop_timesteps_with_invalid_sample_interval(ds, measurement_intervals, verbose=True, logger=None):
+    """Drop timesteps with unexpected sample intervals."""
+    # TODO
+    # - correct logged sample_interval for trailing seconds. Example (58,59,61,62) converted to 60 s ?
+    # - Need to know more how Parsivel software computes sample_interval variable ...
+
+    # Retrieve logged sample_interval
+    sample_interval = ds["sample_interval"].compute().data
+    timesteps = ds["time"].compute().data
+    is_valid_sample_interval = np.isin(sample_interval.data, measurement_intervals)
+    indices_invalid_sample_interval = np.where(~is_valid_sample_interval)[0]
+    if len(indices_invalid_sample_interval) > 0:
+        # Log information for each invalid timestep
+        invalid_timesteps = pd.to_datetime(timesteps[indices_invalid_sample_interval]).strftime("%Y-%m-%d %H:%M:%S")
+        invalid_sample_intervals = sample_interval[indices_invalid_sample_interval]
+        for tt, ss in zip(invalid_timesteps, invalid_sample_intervals):
+            msg = f"Unexpected sampling interval ({ss} s) at {tt}. The measurement has been dropped."
+            log_warning(logger=logger, msg=msg, verbose=verbose)
+        # Remove timesteps with invalid sample intervals
[…]
