### 2025-02-10 #156 @ghiggi
Hi @lwasser,

Thank you for following up. A lot of time has passed since we initially requested the software review.
Our original plan was to go through the pyOpenSci review and then publish a manuscript in JOSS. However, the project has since evolved, and we are now preparing publications in two other journals. Given this shift, I think it makes sense to close the process here and allow the limited reviewer time to be allocated to other software reviews ;)
